updaterar mig

+++ src/view/GameView.java
@@ -47,7 +47,7 @@ implements MouseWheelListener, MouseMotionListener, MouseListener, GameObserver,
 {
 
 	//Temp variable until the proper world gets used
-	int worldSize=800;
+	double worldSize=800;
 	
 	// Determines the zoom level
 	protected float zoom = 1;
